wd last

+++ system_programming/wd/wd.c
@@ -1,28 +1,52 @@
-#include <stdio.h>
-#include <pthread.h>
-#include <sys/types.h>
-#include <unistd.h>
-#include <assert.h>
-#include <time.h> /* clock() */
-#include <signal.h> /* SIGUSR2, SIGTERM */
+/************************************
+ *				WD 					*
+ *			Shye Shapira			*
+ ************************************/
 
-#include "/home/codezila/git/system_programming/include/scheduler.h"
-#include "wd_imp.h"
+#include <pthread.h>	/* pthread_create() */
+#include <unistd.h>		/* fork() excel() getppid() */
+#include <assert.h>		/* assert() */
+#include <time.h>		/* clock() */
+#include <signal.h>		/* SIGUSR2, SIGTERM */
 
-wd_t *WDStart(const char *filename, init_status_t *status)
+#include "scheduler.h"	/* scheduler IMP */
+#include "wd_imp.h"		/* wd imp IMP */
+
+sem_t *sem_is_wd_up;
+
+wd_t *WDStart(const char *filename, status_t *status)
 {
 	wd_t *app_wd = NULL;
 	pid_t wd_pid = 0;
+	int sem_val = 0;
 
 	assert(NULL != filename);
 	assert(NULL != status);
 
 	app_wd = WDInit(status);
+	if (NULL == app_wd)
+	{
+		return NULL;
+	}
+
+	app_wd->my_filename = filename;
+	app_wd->partner_filename = WD_PATH;
 	
+	sem_is_wd_up = sem_open("/sem_is_wd_up", O_CREAT, SEM_PERMS, 0);
+	if (SEM_FAILED == sem_is_wd_up)
+	{
+		*status = OS_FUNC_FAIL;
+
+		return NULL;
+	}
+
 	sem_self = sem_open("/sem_app_ready", O_CREAT, SEM_PERMS, 0);
 	if (SEM_FAILED == sem_self)
 	{
-		*status = SEM_FAIL;
+		*status = OS_FUNC_FAIL;
+		
+		sem_close(sem_is_wd_up);
+		sem_unlink("/sem_is_wd_up");
 
 		return NULL;
 	}
@@ -30,30 +54,80 @@ wd_t *WDStart(const char *filename, init_status_t *status)
 	sem_partner = sem_open("/sem_wd_ready", O_CREAT, SEM_PERMS, 0);
 	if (SEM_FAILED == sem_partner)
 	{
-		*status = SEM_FAIL;
+		*status = OS_FUNC_FAIL;
+
+		sem_close(sem_is_wd_up);
+		sem_unlink("/sem_is_wd_up");
+		sem_close(sem_self);
+		sem_unlink("/sem_app_ready");
 
 		return NULL;
 	}
 
 	if (0 != pthread_create(&app_wd->app_thread, NULL, WDSchedulerRun, app_wd))
 	{
-		*status = PTHREAD_CREATE_FAIL;
-		return app_wd;
-	}
+		*status = OS_FUNC_FAIL;
 
-	wd_pid = fork();
-	if (0 > wd_pid)
+		sem_close(sem_is_wd_up);
+		sem_unlink("/sem_is_wd_up");
+		sem_close(sem_self);
+		sem_unlink("/sem_app_ready");
+		sem_close(sem_partner);
+		sem_unlink("/sem_wd_ready");
+	
+		return NULL;
+	}
+	
+	if (0 > sem_getvalue(sem_is_wd_up, &sem_val))
 	{
-		*status = FORK_FAIL;
-		return app_wd;
+		*status = OS_FUNC_FAIL;
+
+		sem_close(sem_is_wd_up);
+		sem_unlink("/sem_is_wd_up");
+		sem_close(sem_self);
+		sem_unlink("/sem_app_ready");
+		sem_close(sem_partner);
+		sem_unlink("/sem_wd_ready");
+
+		return NULL;
 	}
 
-	else if (0 == wd_pid)
+	if (0 == sem_val)
 	{
-		execl("./wd.out", "./wd.out", NULL);
+		wd_pid = fork();
+		if (0 > wd_pid)
+		{
+			*status = OS_FUNC_FAIL;
+
+			sem_close(sem_is_wd_up);
+			sem_unlink("/sem_is_wd_up");
+			sem_close(sem_self);
+			sem_unlink("/sem_app_ready");
+			sem_close(sem_partner);
+			sem_unlink("/sem_wd_ready");
+
+			return app_wd;
+		}
+
+		else if (0 == wd_pid)
+		{
+			execl(WD_PATH, WD_PATH, filename, NULL);
+		}
+
+		partner_pid = wd_pid;
+
+		if (0 > sem_post(sem_is_wd_up))
+		{
+			*status = OS_FUNC_FAIL;
+			
+			return NULL;
+		}
 	}
 
-	partner_pid = wd_pid;
+	else
+	{
+		partner_pid = getppid();
+	}
 	
 	return app_wd;
 }
@@ -66,21 +140,48 @@ void WDStop(wd_t *app_wd)
 	assert(NULL != app_wd);
 
 	while (0 == sem_val &&
-			1 > ((clock() - start) / CLOCKS_PER_SEC))
+		   1 > ((clock() - start) / CLOCKS_PER_SEC))
 	{
 		kill(partner_pid, SIGUSR2);
 		if (0 > sem_getvalue(sem_wd_received_stop_signal, &sem_val))
 		{
+			app_wd->status = OS_FUNC_FAIL;
 
+			return;
 		}
 	}
 
 	if (0 == sem_val)
 	{
-		sem_post(sem_wd_received_stop_signal);
+		if (0 != sem_post(sem_wd_received_stop_signal))
+		{
+			app_wd->status = OS_FUNC_FAIL;
+
+			return;
+		}
+
 		kill(partner_pid, SIGTERM);
 	}
 
-	pthread_join(app_wd->app_thread, NULL);
+	if (0 != sem_close(sem_is_wd_up))
+	{
+		app_wd->status = OS_FUNC_FAIL;
+
+		return;
+	}
+
+	if (0 != sem_unlink("/sem_is_wd_up"))
+	{
+		app_wd->status = OS_FUNC_FAIL;
+
+		return;
+	}
+
+	if (0 != pthread_join(app_wd->app_thread, NULL))
+	{
+		app_wd->status = OS_FUNC_FAIL;
+
+		return;
+	}
 }
